requirements.txt
# Core dependencies
pyyaml>=6.0
astroid>=3.0.0
tree-sitter>=0.20.0
tree-sitter-languages>=1.7.0
colorama>=0.4.6
rich>=13.0.0

# Development dependencies
pytest>=7.0.0
black>=23.0.0